fix: resolve circular dependencies causing runtime initialization error
Extract DialogBackdrop/DialogContainer to dialog-shared.tsx and
ProfileSetupData to profile-setup.types.ts to break 7 circular
dependency cycles that caused "Cannot access 'ft' before initialization"
in the production bundle.
