+++ src/components/profile/setup/ProfileSetupStep3Social.tsx
@@ -2,7 +2,7 @@ import { motion } from "@/lib/motion";
 import { Instagram, Globe, Music2 } from "@/lib/icons";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import type { ProfileSetupData } from "./EnhancedProfileSetup";
+import type { ProfileSetupData } from "./profile-setup.types";
 import type { SocialLinks } from "@/types/profile";
 
 interface ProfileSetupStep3SocialProps {
